+++ device-bootloaders/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-5479] Reporting trace buffer threshold event to host
 - [SW-15176] Preserve flash config space with firmware update
 - [SW-15176] Separated persistant and non persistant configration structs in BL2 flash config data
+- [SW-15559] Adding a stability check loop for pmic set voltage function
 ### Changed
 - Fixing MM reset boot flow.
 - [SW-13350] Replacing Trace_String() with Trace_Format_String_V()